==> src/02-nettoyage.ipynb <==
# Import des outils / jeu de données

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

from constants import var_categoriques_original as var_categoriques
from constants import var_numeriques

np.random.seed(0)
sns.set_theme()

df = pd.read_csv("../data/data.csv", sep=",", index_col="ID", parse_dates=True)

df.head()

## Variables globales

df[var_categoriques] = df[var_categoriques].astype(str).astype("category")

**Tableau.** Tâches effectuées durant le nettoyage des données

| Type                                 | Description                                                                                                         |
|:-------------------------------------|:--------------------------------------------------------------------------------------------------------------------|
| **Données parasites**                | Réponses extravagantes au questionnaire<br>Variables constantes et non-renseignées (`Z_CostContact` et `Z_Revenue`) |
| **Données dupliquées**               | 8% d'individus en double (_182 individus_)                                                                          |
| **Variables pas assez informatives** | Variable catégorique avec plus de 99% d'individus dans une seule classe                                             |
| **Standardisation**                  | Modalités de variable catégoriques en double (_"Seul" / "Célibataire"_)                                             |
| **Valeurs extrêmes**                 | Date de naissance trop anciennes (_1800_)<br>Revenu non crédible (_666 666$_)                                       |
| **Valeurs manquantes**               | Revenu manquant pour 1% des individus => remplacement par la moyenne                                                |

# Nettoyage des données

## Données parasites


### Variables

Dans le jeu de données, 2 colonnes sont non documentées et contiennent des valeurs constantes :
- `Z_CostContact`
- `Z_Revenue`

**Variable `Z_CostContact`**

print(
    f'Il y a {np.sum(df["Z_CostContact"] == 3)} / {len(df)} individus pour lesquelles la variable Z_CostContact vaut 3.'
)

df["Z_CostContact"].hist()

**Variable `Z_Revenue`**

print(
    f'Il y a {np.sum(df["Z_Revenue"] == 11)} / {len(df)} individus pour lesquelles la variable Z_Revenue vaut 11.'
)

df["Z_Revenue"].hist()

Nous décidons de les supprimer, car elles sont parasites et ne fournissent pas d'information.

df.drop(columns=["Z_CostContact", "Z_Revenue"], inplace=True)

### Valeurs

#### Statut marital

sns.histplot(y=df["Marital_Status"])

Parmi les situations maritales, on observe "absurde" et "YOLO" (You Only Live Once, _on ne vit qu'une fois_).

Comme on ne peut pas déduire ce que veut dire "absurde" dans ce contexte, on se décide de supprimer l'individu.
De même, on suppose que la personne ayant répondu "YOLO", donc ayant répondu de façon humoristique, doit être supprimée du jeu de données.

df.drop(index=df.loc[df["Marital_Status"] == "YOLO"].index, inplace=True)
df.drop(index=df.loc[df["Marital_Status"] == "Absurd"].index, inplace=True)

# On retire les options "YOLO" et "Absurd" de nos variables catégoriques.
df["Marital_Status"] = df["Marital_Status"].cat.remove_categories(["YOLO", "Absurd"])

sns.histplot(y=df["Marital_Status"])

## Données dupliquées

### Individus

sum(df.duplicated())

Nous observons qu'il y a 182 lignes qui sont dupliquées.

Nous pouvons supposer que les données ont été fusionnées à partir de plusieurs sources, et que ces sources avaient des individus en commun : nous décidons donc de supprimer les doublons.

df.drop_duplicates(inplace=True)

sum(df.duplicated())

## Variables pas assez informatives

# todo: dire que la variable "Complain" (si le client s'est plaint ou non), n'apporte pas assez d'information pour être gardée, car il y a moins de 1% des observations qui ont une classe autre que "0", et qu'en plus ce n'est pas central à notre étude

# todo: essayer quand même une visualisation des individus s'étant plaint par rapport à nos variables cibles, au cas où...

sns.histplot(df["Complain"], shrink=0.5)

print(
    f'Il y a {100 * df["Complain"].astype(int).sum() / len(df["Complain"]):.2f}% des individus qui se sont plaints.'
)

df.drop(columns=["Complain"], inplace=True)

## Standardisation

### Modalités de variables catégoriques

sns.histplot(y=df["Marital_Status"])

Sans plus d'information, "Alone" est une autre formulation de "Single", on décide donc de remplacer "Alone" par "Single".

df["Marital_Status"].replace("Alone", "Single", inplace=True)

sns.histplot(y=df["Marital_Status"])

## Valeurs extrêmes

### Income

_, ax = plt.subplots(1, 2, figsize=(10, 3))
sns.boxplot(df["Income"], width=0.25, ax=ax[0])
sns.histplot(df["Income"], kde=True, ax=ax[1])
plt.show()

# todo: texte pour commenter

df["Income"][df["Income"] > 200_000]

Pour nous, il s'agit d'une valeur aberrante à supprimer du jeu de données car "666 666" ne semble pas être une valeur réelle (6 fois le nombre 6...).

df.drop(index=df[df["Income"] > 200_000].index, inplace=True)

_, ax = plt.subplots(1, 2, figsize=(10, 3))
sns.boxplot(df["Income"], width=0.25, ax=ax[0])
sns.histplot(df["Income"], kde=True, ax=ax[1])
plt.show()

### Year Date

_, ax = plt.subplots(1, 2, figsize=(10, 3))
sns.boxplot(df["Year_Birth"], width=0.25, ax=ax[0])
sns.histplot(df["Year_Birth"], kde=True, ax=ax[1])
plt.show()

df[df["Year_Birth"] < 1920]

Comme les personnes les plus âgées au monde ont très rarement plus de 100 ans, et que nous sommes en 2023, on peut sans problème estimer que les trois dates de naissance avant 1920 sont erronées.

Pour les deux années 1893 et 1899, on peut s'imaginer que les personnes ont fait une faute de frappe, et voulaient écrire, respectivement, 1983 et 1989.

df.loc[df["Year_Birth"] == 1893, "Year_Birth"] = 1983
df.loc[df["Year_Birth"] == 1899, "Year_Birth"] = 1989

Pour l'année 1900, il est difficile d'imaginer une faute de frappe, il s'agit peut-être de la valeur par défaut.
On peut tout simplement supprimer la ligne correspondante.

df.drop(index=df.loc[df["Year_Birth"] == 1900, "Year_Birth"].index, inplace=True)

_, ax = plt.subplots(1, 2, figsize=(10, 3))
sns.boxplot(df["Year_Birth"], width=0.25, ax=ax[0])
sns.histplot(df["Year_Birth"], kde=True, ax=ax[1])
plt.show()

## Valeurs manquantes

sns.heatmap(
    df.isna(), cbar=False
)  # todo: mettre les graphiques à côté (horizontalement)

sns.displot(
    data=df.isna().melt(value_name="missing"),
    y="variable",
    hue="missing",
    multiple="fill",
    aspect=1.25,
)

plt.title("Nombre de valeurs manquantes pour chaque variable")
df.isna().sum().plot.barh()

# todo : commenter

# TODO: supprimer valeurs manquantes OU remplacer par la moyenne ?

# todo: expliquer POURQUOI ces données sont manquantes (notre hypothèse du moins), et expliquer notre démarche en conséquence

### Remplacement par la moyenne

df["Income"][df["Income"].isna()]

df["Income"].fillna(df["Income"].mean(), inplace=True)

# Vérification du bon remplacement des valeurs manquantes
df["Income"].isna().sum()

# Sauvegarde du Dataframe

df.to_csv("../data/data-cleaned.csv")
