==> src/08b-bilan-regression.ipynb <==
# Import des outils / jeu de données

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

np.random.seed(0)
sns.set_theme()

# Bilan des régressions

score_modeles = pd.read_csv("../data/results/regressions.csv")

sns.barplot(data=score_modeles, y="Modèle", x="Valeur", hue="Métrique")

score_modeles = score_modeles[~(score_modeles["Modèle"]).isin(["Réseau de Neurones"])]

plt.title("Prédiction")
sns.barplot(data=score_modeles, y="Modèle", x="Valeur", hue="Métrique")
